Refactor `ZipImpl` by extracting logic into `AbstractZip`
